Update lib/protocol/packets/RowDataPacket.js
Added Z to end of dateString to force using UTC time for database, as
this solves the problem for our project. Otherwise dateString is
always evaluated at the process' localtime. That is a valid solution
if the process is in the same tz as the database and all databases
that the process is connecting to are in this same tz. The proposed
solution is to set your process to the tz of the database. The correct
solution would be to be able to set the tz uniquely for each database
(this has been proposed by some who have forked the project).